+++ src/Test/Util/mmdCameraAutoFocus.ts
@@ -41,7 +41,6 @@ export class MmdCameraAutoFocus {
         const defaultPipeline = this._pipeline;
         const rotationMatrix = new Matrix();
         const cameraNormal = new Vector3();
-        const cameraEyePosition = new Vector3();
         const skeletonWorldMatrix = this._skeletonWorldMatrix;
         const boneWorldMatrix = new Matrix();
         const headRelativePosition = new Vector3();
@@ -58,16 +57,10 @@ export class MmdCameraAutoFocus {
 
             Vector3.TransformNormalFromFloatsToRef(0, 0, 1, rotationMatrix, cameraNormal);
 
-            camera.position.addToRef(
-                Vector3.TransformCoordinatesFromFloatsToRef(0, 0, camera.distance, rotationMatrix, cameraEyePosition),
-                cameraEyePosition
-            );
-
             if (camera.parent !== null) {
                 camera.parent.computeWorldMatrix();
                 const cameraParentWorldMatrix = camera.parent.getWorldMatrix();
 
-                Vector3.TransformCoordinatesToRef(cameraEyePosition, cameraParentWorldMatrix, cameraEyePosition);
                 Vector3.TransformNormalToRef(cameraNormal, cameraParentWorldMatrix, cameraNormal);
                 cameraNormal.normalize();
             }
@@ -79,7 +72,7 @@ export class MmdCameraAutoFocus {
             }
 
             boneWorldMatrix.getTranslationToRef(headRelativePosition)
-                .subtractToRef(cameraEyePosition, headRelativePosition);
+                .subtractToRef(camera.globalPosition, headRelativePosition);
 
             defaultPipeline.depthOfField.focusDistance = (Vector3.Dot(headRelativePosition, cameraNormal) / Vector3.Dot(cameraNormal, cameraNormal)) * 1000;
         };